Refactor and resctruture code for DRY
